requirements.txt
PyQt6
PyMuPDF
openpyxl
python-pptx
python-docx
extract-msg
chardet